+++ src/ts/modules/API.ts
@@ -91,24 +91,6 @@ class API {
     return str
   }
 
-  // 检查给定的字符串解析为数字后，是否大于 0
-  static checkNumberGreater0(arg: string) {
-    let num = parseInt(arg)
-    // 空值会是 NaN
-    if (!isNaN(num) && num > 0) {
-      // 符合条件
-      return {
-        result: true,
-        value: num,
-      }
-    }
-    // 不符合条件
-    return {
-      result: false,
-      value: 0,
-    }
-  }
-
   // 从 url 中获取指定的查询字段的值
   // 注意：返回值经过 encodeURIComponent 编码！
   static getURLSearchField(url: string, query: string) {